Move integration tests to its own file
Since the integration tests in the `FileWatcher` tests are actually
testing 3 different classes (and interacting with the OS), we move it
to its own file, and we call it more generally.

We also mark the tests with the pytest mark `integration` to make it
more clear they are integration tests, and to enable an easy way to
only run (exclude) integration tests.
