+++ pyproject.toml
@@ -74,6 +74,9 @@ src_paths = ["src", "examples", "tests"]
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 required_plugins = ["pytest-asyncio", "pytest-mock"]
+markers = [
+  "integration: integration tests (deselect with '-m \"not integration\"')",
+]
 
 [[tool.mypy.overrides]]
 module = ["async_solipsism", "async_solipsism.*"]
